Updated docs generation script for Linux platform.

+++ docs/generate_docs.sh
@@ -2,10 +2,10 @@
 ## Must be run from within the venv.
 
 # Usage:
-# ./generate_docs.sh - to just regenerate the main docs
+# ./generate_docs.sh - to just regenerate sphinx docs
 # ./generate_docs.sh clean - to clean all the build files
-# ./generate_docs.sh rebuild - to rebuild the Doxygen definitions
-# ./generate_docs.sh clean rebuild - to clean the build files and rebuild the definitions
+# ./generate_docs.sh rebuild - to rebuild the sphinx with new build files
+# ./generate_docs.sh clean rebuild - to clean the build files and rebuild with new build files
 
 set -e
 
@@ -14,11 +14,19 @@ if [[ ! -d "_build" || ! -d "apidoc" || ! -d "../build/xml" ]]; then
 fi
 
 if [[ $1 == "clean" ]]; then
-    echo "Deleting _build folder"
+    echo "Deleting sphinx _build folder"
     rm -rf _build
 
-    echo "Deleting apidoc folder"
-    rm -rf apidoc
+    echo "Deleting API reference pages"
+    rm -rf doxygen/api_member_list/class
+    rm -rf doxygen/api_member_list/file
+    rm -rf doxygen/api_member_list/struct
+    rm doxygen/api_member_list/classlist.rst
+    rm doxygen/api_member_list/filelist.rst
+    rm doxygen/api_member_list/structlist.rst    
+
+    echo "Deleting doxygen manual pages"
+    python dox_converter.py --del
 
     shift
     if [[ $1 == "" ]]; then
@@ -26,19 +34,14 @@ if [[ $1 == "clean" ]]; then
     fi
 fi
 
-# if [[ $1 == "rebuild" ]]; then
-#     cmake -S .. -B ../build
-#     make --no-print-directory -C ../build -j"$(nproc)" doxygen
-#     breathe-apidoc --generate class --members --force --output-dir apidoc ../build/docs/doxygen/xml/
-#     python add_undoc_members.py
-# fi
-
-breathe-apidoc --generate class --members --force --output-dir doxygen/api_member_list doxygen/output/xml/
-breathe-apidoc --generate file --force --output-dir doxygen/api_member_list doxygen/output/xml/
-breathe-apidoc --generate struct --members --force --output-dir doxygen/api_member_list doxygen/output/xml/
-
-rm doxygen/api_member_list/file/*dox.rst  # Removing redundant doxygen manual pages that were converted to rst pages to avoid sphinx errors
-
-python dox_converter.py
+if [[ $1 == "rebuild" ]]; then
+    cmake -S .. -B ../build
+    cmake --build ../build -- doxygen
+    breathe-apidoc --generate class --members --force --output-dir doxygen/api_member_list ../build/docs/doxygen/xml/
+    breathe-apidoc --generate file --force --output-dir doxygen/api_member_list ../build/docs/doxygen/xml/
+    breathe-apidoc --generate struct --members --force --output-dir doxygen/api_member_list ../build/docs/doxygen/xml/
+    rm doxygen/api_member_list/file/*dox.rst
+    python dox_converter.py
+fi
 
 make -j"$(nproc)" html
